Use separate header files
I've broken apart the monolithic hst.h header file into a bunch of
separate headers — basically one for each source file.
