+++ tests/test-operators.c
@@ -5,9 +5,12 @@
  * -----------------------------------------------------------------------------
  */
 
-#include <string.h>
+#include "operators.h"
 
-#include "hst.h"
+#include "basics.h"
+#include "behavior.h"
+#include "environment.h"
+#include "id-set.h"
 #include "test-case-harness.h"
 #include "test-cases.h"
